requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
memorisdk==0.1.85
openai==1.51.2
anthropic==0.39.0
google-generativeai==0.8.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
requests==2.32.3
python-multipart==0.0.12
